+++ apps/web/app/routes/params.$id.tsx
@@ -1,4 +1,4 @@
-import type { MetaFunction } from "@remix-run/cloudflare";
+import type { MetaFunction } from "@react-router/cloudflare";
 import { UserDetailScreen } from "app/features/user/detail-screen";
 
 export const meta: MetaFunction = () => {