+++ apps/web/app/routes/sign-up.tsx
@@ -1,4 +1,4 @@
-import type { MetaFunction } from "@remix-run/cloudflare";
+import type { MetaFunction } from "@react-router/cloudflare";
 import { SignUpScreen } from "app/features/sign-up/screen";
 
 export const meta: MetaFunction = () => {