+++ apps/web/app/routes/virtualized-list.tsx
@@ -1,4 +1,4 @@
-import type { MetaFunction } from "@remix-run/cloudflare";
+import type { MetaFunction } from "@react-router/cloudflare";
 import { VirtualizedListScreen } from "app/features/virtualized-list/screen";
 
 export const meta: MetaFunction = () => {