+++ apps/web/worker.ts
@@ -1,12 +1,8 @@
-import { createRequestHandler, logDevReady } from "@remix-run/cloudflare";
+import { createRequestHandler } from "@react-router/cloudflare";
 import { getLoadContext, type Env } from "./load-context";
 
 // @ts-ignore - This will be generated after build
 import * as build from "./build/server";
-
-if (process.env.NODE_ENV === "development") {
-  logDevReady(build);
-}
 
 export default {
   async fetch(request: Request, env: Env, ctx: ExecutionContext) {